BU-316
Changed relationship of CUs and speleothems:
There can be multiple CUs in a speleothem but a CU can only be linked to one speleothem.
For next version of R package, it may be worthy to do a subclass of Site and assign the slot speleothem in Site rather than in CU because this way, we can portray that a speleothem can contain multiple CUs as per the data model.